requirements.txt
flask
python-dotenv
requests
Authlib
mysql-connector-python